backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
PyMuPDF==1.23.21
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.1
